FM: don't destroy notification channel on stop
There's no reason for the notification channel to be destroyed on FM
stop and recreated on every app launch.

This also crashes the app because the notification belonging to this
channel is still being displayed.

Even if the intention behind the method would be to remove the actual
notification, this is not necessary as this will happen automatically
when calling stopForeground(true).

Change-Id: Id84d247733edb66fe8755f2fe4625e97066d4958